requirements.txt
requests
pymongo
graphene
graphene_mongo
flask
bcrypt
flask_graphql
python-dotenv
faker
coverage
